// frontend/src/app/core/services/api/model/userBatchMigrationRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserBatchMigrationRequestItem } from './userBatchMigrationRequestItem';


/**
 * Request payload for batch migrating users from Excel
 */
export interface UserBatchMigrationRequest { 
    users: Array<UserBatchMigrationRequestItem>;
}