// frontend/src/app/core/services/api/model/userBatchMigrationRequestItem.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SystemRole } from './systemRole';


/**
 * Single user item within a batch migration request
 */
export interface UserBatchMigrationRequestItem { 
    /**
     * First name of the user
     */
    firstName: string;
    /**
     * Last name of the user
     */
    lastName: string;
    /**
     * Date of birth
     */
    dateOfBirth: string;
    /**
     * Email address (also used as username if username not provided)
     */
    email: string;
    /**
     * Username for login (optional, defaults to email)
     */
    username?: string;
    role: SystemRole;
}
export namespace UserBatchMigrationRequestItem {
}