// frontend/src/app/core/services/api/model/userBatchMigrationResponse.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';
import { UserBatchMigrationSkippedItem } from './userBatchMigrationSkippedItem';


/**
 * Response from a batch user migration
 */
export interface UserBatchMigrationResponse { 
    /**
     * Successfully created users
     */
    created?: Array<UserDto>;
    /**
     * Users that were skipped (e.g., email already exists)
     */
    skipped?: Array<UserBatchMigrationSkippedItem>;
}